// src/generated/contentful-types.ts
export interface Page {
  sys: {
    id: string;
    publishedAt?: string;
    updatedAt?: string;
    revision?: number;
  };
  name: string;
  title: string;
  slug: string;
  contentViewCollection?: { items: Array<ContentView> };
  imageCollection?: { items: Array<{ sys: { id: string }; url?: string }> };
  description?: string;
  sharingTitle?: string;
  sharingImage?: { sys: { id: string }; url?: string };
}
// ...
export interface ContentView {
  sys: {
    id: string;
    publishedAt?: string;
    updatedAt?: string;
    revision?: number;
  };
  name: string;
  title?: string;
  type: string;
  dataAdapter: { sys: { id: string } };
  targetCollection?: { items: Array<{ sys: { id: string } }> };
  isAtomic: boolean;
  isTitleVisible: boolean;
  keys?: Array<string>;
  values?: Array<string>;
}